# Devano/PT MOB openAPI.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "3000",
        "title": "PT MOB json",
        "license": {
            "name": "MIT"
        }
    },
    "servers": [
        {
            "url": "localhost:3000"
        }
    ],
    "paths": {
        "/author": {
            "get": {
                "summary": "Get list and search of author",
                "tags": [
                    "author"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id of author",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fullname",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return list of Author",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListOfAuthor"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "create author object",
                "requestBody": {
                    "description": "json of author object to create",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Author"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "return saved author object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Author"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/author/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            ],
            "get": {
                "description": "get detail of author by id",
                "responses": {
                    "200": {
                        "description": "return  detail of author",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Author"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "update",
                "requestBody": {
                    "description": "json object of author to update",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Author"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "return  updated object of author",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Author"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "delete",
                "responses": {
                    "200": {
                        "description": "return list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ListOfAuthor": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/Author"
                }
            },
            "Author": {
                "type": "object",
                "required": [
                    "id",
                    "email",
                    "fullname"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "fullname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                }
            },
...
            "Message": {
                "type": "object",
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    }
                }
            }
        },
        "securitySchemes": {
            "BasicAuth": {
                "type": "http",
                "scheme": "basic"
            }
        }
    },
    "security": [
        {
            "BasicAuth": []
        }
    ]
}